+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [x]
 * Not relevant status
+* Notifications4Plugin library
+* PHP7Backport for libraries
 
 ## [1.2.1]
 * Adjustments to work with latest Version of Notifications4Plugins