+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [x]
 * Not relevant status
 * Absent / Total column
+* Calc auto minimum_attendance
 * Notifications4Plugin library
 * PHP7Backport for libraries