+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [x]
 * Notifications4Plugin library
 * PHP7Backport for libraries
+* Fix PHP 7.2
 
 ## [1.2.1]
 * Adjustments to work with latest Version of Notifications4Plugins